// packages/core/src/utils/clients/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      NodeVersion: {
        Row: {
          created_at: string
          description: string | null
          extras: Json
          handoffs: string[] | null
          llm_model: string
          memory: Json | null
          node_id: string
          system_prompt: string
          tools: string[]
          updated_at: string
          version: number
          waiting_for: string[] | null
          wf_version_id: string
        }
        Insert: {
          created_at?: string
          description?: string | null
          extras: Json
          handoffs?: string[] | null
          llm_model: string
          memory?: Json | null
          node_id?: string
          system_prompt: string
          tools: string[]
          updated_at?: string
          version: number
          waiting_for?: string[] | null
          wf_version_id: string
        }
        Update: {
          created_at?: string
          description?: string | null
          extras?: Json
          handoffs?: string[] | null
          llm_model?: string
          memory?: Json | null
          node_id?: string
          system_prompt?: string
          tools?: string[]
          updated_at?: string
          version?: number
          waiting_for?: string[] | null
          wf_version_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "NodeVersion_wf_version_id_fkey"
            columns: ["wf_version_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never